Fix MangaDex’s followStatus & type of TrackStatus’s property

+++ app/src/main/java/eu/kanade/domain/track/interactor/TrackChapter.kt
@@ -34,7 +34,7 @@ class TrackChapter(
                     service == null ||
                     !service.isLoggedIn ||
                     chapterNumber <= track.lastChapterRead /* SY --> */ ||
-                    (service is MdList && track.status == FollowStatus.UNFOLLOWED.long.toLong())/* SY <-- */
+                    (service is MdList && track.status == FollowStatus.UNFOLLOWED.long)/* SY <-- */
                 ) {
                     return@mapNotNull null
                 }
